+++ src/typescript/ui/UnmuteElement.ts
@@ -8,8 +8,6 @@ import {EventType} from "../events/EventType";
  */
 export class UnmuteElement extends GraphicElement {
 
-    private wasUnmuted:boolean = false;
-
     /**
      * Constructor
      * @param stormPlayer reference to the main player class
@@ -73,9 +71,8 @@ export class UnmuteElement extends GraphicElement {
 
         this.stormPlayer.addEventListener(EventType.LIBRARY_INITIALIZED, function () {
             that.stormPlayer.getLibrary().addEventListener("volumeChange", function (event: any) {
-                if (event.isMuted && that.wasUnmuted == false) {
+                if (event.isMuted && event.type == "browser") {
                     that.show();
-                    that.wasUnmuted = true;
                 } else
                     that.hide();
             });